### 2024-02-15 #6355 @danielskatz
Note: authors' note with submission includes:

>The EdgeVPNio (aka Evio) is based on the work "EdgeVPN: Self-organizing layer-2 virtual edge networks", previously published in the journal Future Generation Computer Systems Volume 140, https://www.sciencedirect.com/science/article/pii/S0167739X22003235/. Succinctly, Evio is a software-defined peer VPN that enables virtual private Ethernet networks that span edge and cloud resources. To demonstrate Evio's functionality, 2 or more Ubuntu 22.04 systems and credentials on a bootstrapping XMPP server are required. The authors have hosted accounts on their development infrastructure for the reviewers' convenience. See the encrypted [...]. Decrypt with [...] in the same repo folder as the paper. Detailed instructions are available at https://edgevpn.io/trial/.
