### 2024-02-15 #6355 @danielskatz
👋 @kcratie - thanks for your submission.

In addition to the paper generation issue, I have a number of concerns:

1. JOSS is for open source software. No one should need a password to access the software or use it.

2. The FGCS paper was published between 11 and 18 months ago, depending on if you count acceptance or print publication. Given this, I'm unsure about how another publication at this point makes sense.

Can you explain your thinking on these two issues?
