### 2024-02-15 #6355 @danielskatz
Regarding the paper compilation error, please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. In this case, you are not correctly defining affiliations in the paper header. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to try to make a new PDF. editorialbot commands need to be the first entry in a new comment.
