### 2024-02-19 #6355 @kcratie
@danielskatz - Thanks for your feedback. In response to your concerns, (1) The software source code and release package are both open source and freely available. Running Evio requires external services for authentication and signaling, and there are open-source solutions for this, e.g., Openfire or ejabberd. Anyone can host their own and independently run Evio. We have hosted this service as a convenience to the JOSS reviewers for demonstrating Evio, and the encrypted archive contains the configuration files for accessing our hosted services. Instructions for doing this yourself are available [here](https://edgevpn.io/documentation/).
(2) The journal publication focuses on the theoretical aspects, design, and evaluation of Evio. Our intent for the JOSS publication is to describe the most recent version of the software implementation, its capabilities, and potential use cases.

> 👋 @kcratie - thanks for your submission.
>
> In addition to the paper generation issue, I have a number of concerns:
>
> 1. JOSS is for open source software. No one should need a password to access the software or use it.
> 2. The FGCS paper was published between 11 and 18 months ago, depending on if you count acceptance or print publication. Given this, I'm unsure about how another publication at this point makes sense.
>
> Can you explain your thinking on these two issues?
