### 2024-02-19 #6355 @danielskatz
Thanks @kcratie - I'm going to ask the JOSS editors to review this to check if this is suitable for JOSS review. You should hear back in a week or two
