### 2024-03-25 #6355 @sneakers-the-rat
Aha sorry, i had mixed up pyopensci with JOSS, where we do have additional editorial checks before seeking reviewers, my mistake!

I have reached out to a few p2p networking people privately, but they said to look for others and they may be available as alternates. Just posted on the fediverse seeking reviewers there.

Taking a random sample of the reviewer database for people with 'distributed systems' in their keywords, i see @arnavdas88 , @achanda , @pradeeban , @TA3 , and @imw as potential reviewers - are any of you interested?
