### 2024-04-17 #6638 @editorialbot
[REVIEW]: EdgeVPN.io: Seamless Software-defined Layer 2 Virtual Networking for Edge Computing
**Submitting author:** @kcratie (<a href="http://orcid.org/0000-0001-8248-2856">Kensworth Subratie</a>)
**Repository:** https://github.com/EdgeVPNio/evio
**Branch with paper.md** (empty if default branch): joss
**Version:** v24.1.2.1061
