### 2024-04-17 #6638 @sneakers-the-rat
Hello @abhishektiwari and @pradeeban, thank you for agreeing to review this package!

As the top issue says, you'll be working from a checklist, a [set of guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and some [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).

## Goals of Review

The goal here is to collaboratively help this package reach a basic standard for usability, maintainability, and correctness - we are all friends here, there are no gates to be kept, we're just trying to learn from and help each other out. Feel free to talk amongst each other as peers in the issues related to this review. This is an open and iterative review, so you don't need to wait until you've reviewed everything before raising an issue or commenting - in fact its preferable to raise a comment early so that the authors have time to respond and make any requested changes while you're completing the rest of your review.

The checklist is a set of minimum standards that the package has to meet, but you are also free to read, comment, raise issues and pull requests, and do whatever else falls under your purview as a reviewer. Think about what you would want to have if you were a potential new user of the package - is the documentation clear? do the tests cover what they need to? etc. That kind of neutral third-party perspective is invaluable for making software accessible and maintainable.

## Process

As you review, please
- Raise issues in the repository and
- link back to this issue so that we can track issues related to this review.

Since github issues don't have threading, this is how we can keep discussion organized, because commenting in this issue can be a bit messy. Don't be shy commenting here to ask questions about the review itself, though!

It's up to you all and the maintainers if you want to follow any sort of naming convention for issues/PRs - authors please comment here about any preferences you have for raised issues. I would recommend at least tagging issues by general topic areas - `[docs]`, `[tests]`, `[bug]`, etc. and also if the reviewers would indicate if a review is a `[question]`, `[suggestion]` or a `[blocker]` for the review so the authors know if a given issue needs some fix in order for the review to be completed.

Other than that, i'll be here so feel free to ask questions! The first thing you'll want to do is generate your checklists using the commands in the OP comment
