### 2024-04-28 #6638 @abhishektiwari
@kcratie Please see my initial comments,

1. Can you clarify what repositories under `https://github.com/EdgeVPNio` are under the scope for this paper? Repository linked to this submission is [evio](https://github.com/EdgeVPNio/evio). Paper itself mentions two repositories [evio](https://github.com/EdgeVPNio/evio) and [tincan](https://github.com/EdgeVPNio/tincan), but there are other complimentary repositories like [portal](https://github.com/EdgeVPNio/portal), [tools](https://github.com/EdgeVPNio/tools), etc.

2. Also can you provide temporary XMPP/TURN/STUN accounts to reviewers to help us bootstrap Evio overlay? I would avoid my spending my time and money to install and run XMPP/TURN/STUN on AWS just to verify the software claims.

Tagging @sneakers-the-rat for visibility.
