### 2024-04-28 EdgeVPNio/evio#32 @pradeeban
Automated tests for evio
This is part of https://github.com/openjournals/joss-reviews/issues/6638 review.

While automated tests are not mandatory and manual steps are okay as part of the JOSS review ("Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"), given the complex nature of the project, automated tests to confirm the framework installs and works (through a set of test cases) will be a great addition.

Is this possible to be incorporated (or is this something that already exists and I did not look careful enough)?
