### 2024-04-30 #6638 @renatof
Thanks, folks!
Regarding temporary XMPP/TURN accounts, we had provided a zip file as part of the submission, but you can also request trial accounts using the form linked from here:
https://edgevpn.io/trial/
