### 2024-04-30 #6638 @renatof
> @kcratie Please see my initial comments,
>
> 1. Can you clarify what repositories under `https://github.com/EdgeVPNio` are under the scope for this paper? Repository linked to this submission is [evio](https://github.com/EdgeVPNio/evio). Paper itself mentions two repositories [evio](https://github.com/EdgeVPNio/evio) and [tincan](https://github.com/EdgeVPNio/tincan), but there are other complimentary repositories like [portal](https://github.com/EdgeVPNio/portal), [tools](https://github.com/EdgeVPNio/tools), etc.
> 2. Also can you provide temporary XMPP/TURN/STUN accounts to reviewers to help us bootstrap Evio overlay? I would avoid spending my time and money to install and run XMPP/TURN/STUN on AWS just to verify the software claims.
>
> Tagging @sneakers-the-rat for visibility.

Folks, our answers below:
1. confirming that the only repositories under the scope of the paper are evio and tincan
2. attaching a zip file here [JOS1.zip](https://github.com/openjournals/joss-reviews/files/15171435/JOS1.zip) with 5 configuration files for trial accounts that can be used for testing as per requested.
Instructions for running Evio with Docker using these accounts are available [here](https://edgevpn.io/trial/)
