### 2024-04-30 EdgeVPNio/evio#32 @kcratie
The software source code and release package are both open source and freely available in the evio and tincan repositories. Running Evio requires external services for authentication and signaling, and we have hosted this service as a convenience to the JOSS reviewers for demonstrating Evio. An archive containing the configuration files for accessing our hosted services has been attached to our review thread #6638. Instructions for getting started with EdgeVPN are available [here](https://edgevpn.io/trial/).
