### 2024-04-30 #6638 @sneakers-the-rat
Thank you @renatof - reviewers please let me know if there are still difficulties. If there is some lack of clarity in the documentation that the reviewers feel like would hinder normal use, they should feel free to describe what difficulties they had and make suggestions to improve onboarding/make it more obvious how to use the tool. Thats part of the goal of the review! To give an external perspective on things that might be hard for people close to the project to see as potential barriers.
