### 2024-05-24 #6638 @renatof
Thanks for the updates, folks - requiring two computers is indeed a bit more than the typical expectation, and while an ideal scenario for testing would be two separate computers, another possibility is to test on two VMs on the same host.
