### 2024-07-13 #6638 @abhishektiwari
@kcratie

I spent a few hours trying to install and run the software using Ubuntu Multipass to create 2 VMs with `XMPP/TURN/STUN` accounts provided above but no luck using [instructions here](https://edgevpn.io/trial/). Is it possible that `XMPP/TURN/STUN accounts` accounts and trial overlay shared with us are expired?

On VMs/containers I just don't see any network connectivity on ping to `trial.edgevpn.io`. It seems overlay has been down for sometime.

```
ping trial.edgevpn.io
PING trial.edgevpn.io (54.71.171.66): 56 data bytes
Request timeout for icmp_seq 0
Request timeout for icmp_seq 1
Request timeout for icmp_seq 2
Request timeout for icmp_seq 3
Request timeout for icmp_seq 4
```

Also provide more detailed instruction here on how to [activate Open vSwitch module](https://edgevpn.io/trial/).

> Ubuntu 22.04 you may need to install extra kernel modules and/or ensure Open vSwitch module is enabled.

@sneakers-the-rat I this point I have spent too much of my time trying to install and run the software but author's have made very difficult to even run and test without spending a few days of effort. Unless authors provide a fully tested step-by-step guide, I don't know how any reviewer can work with half baked instructions to validate functional claims of the software.
