### 2024-07-13 #6638 @abhishektiwari
@kcratie Latest update.

Ok, after re-installing everything from scratch then activating Open vSwitch module on Ubuntu 22.04. I do see node ping working as described.

```
ubuntu@primary:~/.evio$ ping 10.10.100.2
PING 10.10.100.2 (10.10.100.2) 56(84) bytes of data.
64 bytes from 10.10.100.2: icmp_seq=1 ttl=64 time=26.1 ms
64 bytes from 10.10.100.2: icmp_seq=2 ttl=64 time=1.69 ms
64 bytes from 10.10.100.2: icmp_seq=3 ttl=64 time=2.10 ms
64 bytes from 10.10.100.2: icmp_seq=4 ttl=64 time=3.59 ms
64 bytes from 10.10.100.2: icmp_seq=5 ttl=64 time=2.35 ms
64 bytes from 10.10.100.2: icmp_seq=6 ttl=64 time=1.27 ms
64 bytes from 10.10.100.2: icmp_seq=7 ttl=64 time=2.56 ms
64 bytes from 10.10.100.2: icmp_seq=8 ttl=64 time=2.24 ms
64 bytes from 10.10.100.2: icmp_seq=9 ttl=64 time=2.54 ms
64 bytes from 10.10.100.2: icmp_seq=10 ttl=64 time=2.08 ms
64 bytes from 10.10.100.2: icmp_seq=11 ttl=64 time=2.43 ms
64 bytes from 10.10.100.2: icmp_seq=12 ttl=64 time=3.39 ms
64 bytes from 10.10.100.2: icmp_seq=13 ttl=64 time=2.30 ms
64 bytes from 10.10.100.2: icmp_seq=14 ttl=64 time=2.59 ms
64 bytes from 10.10.100.2: icmp_seq=15 ttl=64 time=6.06 ms
64 bytes from 10.10.100.2: icmp_seq=16 ttl=64 time=3.45 ms
^C
--- 10.10.100.2 ping statistics ---
16 packets transmitted, 16 received, 0% packet loss, time 15687ms
rtt min/avg/max/mdev = 1.270/4.171/26.125/5.764 ms
```

```
ubuntu@secondry:~/.evio$ ping 10.10.100.1
PING 10.10.100.1 (10.10.100.1) 56(84) bytes of data.
64 bytes from 10.10.100.1: icmp_seq=1 ttl=64 time=2.30 ms
64 bytes from 10.10.100.1: icmp_seq=2 ttl=64 time=2.42 ms
64 bytes from 10.10.100.1: icmp_seq=3 ttl=64 time=3.21 ms
64 bytes from 10.10.100.1: icmp_seq=4 ttl=64 time=2.62 ms
64 bytes from 10.10.100.1: icmp_seq=5 ttl=64 time=2.41 ms
64 bytes from 10.10.100.1: icmp_seq=6 ttl=64 time=2.15 ms
64 bytes from 10.10.100.1: icmp_seq=7 ttl=64 time=4.02 ms
64 bytes from 10.10.100.1: icmp_seq=8 ttl=64 time=2.62 ms
64 bytes from 10.10.100.1: icmp_seq=9 ttl=64 time=2.29 ms
64 bytes from 10.10.100.1: icmp_seq=10 ttl=64 time=1.22 ms
^C
--- 10.10.100.1 ping statistics ---
10 packets transmitted, 10 received, 0% packet loss, time 9364ms
rtt min/avg/max/mdev = 1.220/2.524/4.015/0.684 ms
```

I was also able to validate the Turn is up and running,

```
nmap -p 3478 trial.edgevpn.io
Starting Nmap 7.94 ( https://nmap.org ) at 2024-07-14 08:54 AEST
Nmap scan report for trial.edgevpn.io (54.71.171.66)
Host is up (0.18s latency).
rDNS record for 54.71.171.66: ec2-54-71-171-66.us-west-2.compute.amazonaws.com

PORT     STATE SERVICE
3478/tcp open  stun

Nmap done: 1 IP address (1 host up) scanned in 0.60 seconds
```
