### 2024-07-13 EdgeVPNio/evio#33 @abhishektiwari
[JOSS Review] Provide detailed instructions on how to install and activate Open vSwitch module on Ubuntu 22.04
Provide detailed instructions on how to install and activate Open vSwitch module on Ubuntu 22.04 as [described on the page](https://edgevpn.io/trial/).

> Note - on Ubuntu 22.04 you may need to install extra kernel modules and/or ensure Open vSwitch module is enabled.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6638)
