### 2024-07-13 #6638 @abhishektiwari
@sneakers-the-rat Marked my reviewer checklist completed and recommend accepting this paper subjected to authors updating the installation instructions as mentioned on [ticket here](https://github.com/EdgeVPNio/evio/issues/33).
