### 2024-07-14 #6638 @renatof
@abhishektiwari thank you for your time and effort and reviewing our software - we sincerely appreciate it. We will update the Open vSwitch installation instructions and add a comment to the ticket you opened when it's ready.
