### 2024-07-16 EdgeVPNio/evio#33 @renatof
Dear @abhishektiwari ,

Thank you again for pointing this out.

@kcratie and I have improved/generalized the installation command with (uname -r) and improved the documentation to be explicit about the three steps needed (install, add to /etc/modules, modprobe)

Let us know if you think this is sufficient, or if you have any additional suggestions?

Thanks again for your time and feedback!
