### 2024-08-05 #6638 @renatof
Hi folks - we did update the Ubuntu 22.04 instructions as [per this ticket](https://github.com/EdgeVPNio/evio/issues/33). Please let us know if there's anything else on our end that needs to be done? Thank you!
