### 2024-08-12 #6638 @sneakers-the-rat
If @kcratie or @renatof could handle the steps above re: making a release, archiving, and posting the version number and DOI here, that would be great :)
