### 2024-08-12 #6638 @sneakers-the-rat
could the authors please add a DOI to the Symphony citation ( 10.5555/1251460.1251470 ) and a URL for "the cloud is not enough" (looks like it's this: https://www.usenix.org/conference/hotcloud15/workshop-program/presentation/zhang )
