### 2024-08-15 #6638 @kcratie
EdgeVPN:
    Release: https://github.com/EdgeVPNio/evio/releases/tag/v24.1.2.1061
    Zenodo: https://zenodo.org/records/10655929
    DOI: https://doi.org/10.5281/zenodo.10655929

Symphony: https://dl.acm.org/doi/10.5555/1251460.1251470
The cloud is not enough: https://dl.acm.org/doi/10.5555/2827719.2827740
