### 2024-08-15 #6638 @sneakers-the-rat
great once https://github.com/EdgeVPNio/evio/pull/34 happens i think we'll be good to go :)
