### 2024-08-16 #6638 @sneakers-the-rat
huh, so those are the DOIs listed by the publisher, but they aren't dereferencing through the DOI system. let me ask EICs what to do about that...
