### 2024-08-20 #6638 @sneakers-the-rat
Ok! Sorry for the double-take on the references, I asked what to do with the ACM citations and apparently they are a known problem, so we updated the reference checker: https://github.com/openjournals/buffy/pull/106

Ill make another PR to switch those overs to URLs in a moment (or you can change, whoever gets to it first) and then we should be good to go! Thanks for bearing with me on this little technicality
