### 2024-08-21 ad221a9
minor changes in paper.bib
