+++ joss/paper.bib
@@ -34,7 +34,7 @@
     doi =       {10.17487/RFC3920},
     url =       {https://www.rfc-editor.org/info/rfc3920},
     author =    {Peter Saint-Andre},
-    title =     {{Extensible Messaging and Presence Protocol (XMPP): Core}},
+    title =     {Extensible Messaging and Presence Protocol ({XMPP}): Core},
     pagetotal = 30,
     year =      2004,
     month =     oct,
@@ -81,7 +81,7 @@ keywords = {ethernet switch, flow-based, virtualization}
     doi =       {10.17487/RFC7047},
     url =       {https://www.rfc-editor.org/info/rfc7047},
     author =    {Ben Pfaff and Bruce Davie},
-    title =     {{The Open vSwitch Database Management Protocol}},
+    title =     {The Open {vSwitch} Database Management Protocol},
     pagetotal = 35,
     year =      2013,
     month =     dec,
@@ -95,14 +95,14 @@ keywords = {ethernet switch, flow-based, virtualization}
     doi =       {10.17487/RFC8926},
     url =       {https://www.rfc-editor.org/info/rfc8926},
     author =    {Jesse Gross and Ilango Ganga and T. Sridhar},
-    title =     {{Geneve: Generic Network Virtualization Encapsulation}},
+    title =     {Geneve: Generic Network Virtualization Encapsulation},
     pagetotal = 34,
     year =      2020,
     month =     nov,
 }
 
 @inproceedings{zhang2015cloud,
-  title={The cloud is not enough: Saving $\{$IoT$\}$ from the cloud},
+  title={The cloud is not enough: Saving {IoT} from the cloud},
   author={Zhang, Ben and Mor, Nitesh and Kolb, John and Chan, Douglas S and Lutz, Ken and Allman, Eric and Wawrzynek, John and Lee, Edward and Kubiatowicz, John},
   booktitle={7th USENIX Workshop on Hot Topics in Cloud Computing (HotCloud 15)},
   year={2015},
@@ -117,7 +117,7 @@ keywords = {ethernet switch, flow-based, virtualization}
     doi =       {10.17487/RFC5245},
     url =       {https://www.rfc-editor.org/info/rfc5245},
     author =    {Jonathan Rosenberg},
-    title =     {{Interactive Connectivity Establishment (ICE): A Protocol for Network Address Translator (NAT) Traversal for Offer/Answer Protocols}},
+    title =     {Interactive Connectivity Establishment ({ICE}): A Protocol for Network Address Translator ({NAT}) Traversal for Offer/Answer Protocols},
     pagetotal = 117,
     year =      2010,
     month =     apr,
