### 2024-08-26 5f5c81b
changes for JOSS publication (#36)
* minor changes in paper.md

* minor changes in paper.bib
