### 2024-08-26 #6638 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
