### 2024-08-26 #6638 @danielskatz
The submission looks good now, so once the Zenodo metadata is changed, we can accept and publish
