### 2024-08-26 #6638 @danielskatz
Congratulations to @kcratie (Kensworth Subratie) and @renatof on your publication!!

JOSS is dependent on its volunteers, and we couldn't succeed without you
